Add files via upload
Added two actual TIFF test images to test out the program with. Due to GitHub upload file-size constraints, only these two images are made available here. This is a very simple test case (the specified optimum number of clusters would here obviously be 2) as these images are at the opposite extremes. Simply unzip them into the image file directory of your choice and run the program, pointing it to the image directory (or hardcode it in the MapSorter.py file).